refactor: extract mcp server runtime into tau-tools
