+++ crates/tau-tools/src/lib.rs
@@ -1,2 +1,5 @@
+pub mod mcp_server_runtime;
 pub mod tool_policy_config;
 pub mod tools;
+
+pub use mcp_server_runtime::*;
